Support registries in gzipped tarballs
Starting with Julia 1.7, a registry can also be stored in a depot as a
gzipped tarball. This commit unpacks the tarball into a temporary
directory. This was the simplest way to support this new feature.

This implementation will unpack a fresh copy of the registry on every
call to `reachable_registries`. We could instead cache the temporary
directory for future calls.

There's an open PR for the Tar.jl package that will be of future
interest:

https://github.com/JuliaIO/Tar.jl/pull/95

This would allow us to instead stream the individual files straight
out of the tarball, rather than having to unpack it into a temporary
location.