Use the Pkg.Registry API
As of Julia 1.7, Pkg.Registry provides some of the same capability
implemented in this package. When `VERSION` is at least v1.7, we use
`Pkg.Registry.reachable_registries`, `Pkg.Registry.RegistryInstance`,
and `Pkg.Registry.PkgEntry` instead of the custom versions implemented
in this package. For lower version numbers, we use the custom versions
implemented in this package.

The largest change when using the new capabilities in Pkg.Registry is
that this package's version of `RegistryInstance` stored a dictionary
mapping package names to `PkgEntry`s, but Pkg.Registry's version has a
dictionary mapping `UUID`s to `PkgEntry`s.

A benefit of using the `Pkg.Registry` API is that it already supports
both uncompressed registries as well as registries stored in
compressed tarballs. Thus, this fixes #37.