Use the RegistryInstances package
The Pkg API often changes with each Julia release. Some of the
capability of Pkg had been duplicated in this package, but it lacked
some of the new capabilities of Pkg. With this change, we now leverage
RegistryInstances, which makes it easier for us to support multiple
versions of Julia.

There are a few differences between the old implementation that was in
this package and the implementation provided by RegistryInstances:
- The old `RegistryInstance` had a `pkgs` field containing a
dictionary mapping package names to `PkgEntry`s. The RegistryInstances
version of `RegistryInstance` has a `pkgs` field, too, but it is a
dictionary mapping `UUID`s to `PkgEntry`s.
- The old `PkgEntry` had a `repo` field. The RegistryInstances
version does not. Instead, it has a lazily initialized `info` field,
which should be accessed using `RegistryInstances.registry_info`.

Another benefit of using RegistryInstances is that it supports both
uncompressed registries as well as registries stored in compressed
tarballs. Thus, this fixes #37.