Fix lookup of stdlibs on Julia v1.8
With Julia v1.8, there were Pkg API changes that caused this package
to break. In particular, our method of determining whether a package
is part of the standard library now fails.

With older versions of Julia, `Pkg.Types.stdlibs()` returns a
dictionary mapping `UUID`s to `String`s, which are just the package
names. As of Julia v1.8, `Pkg.Types.stdlibs()` returns a dictionary
mapping `UUID`s to `Tuple{String, Union{Nothing, VersionNumber}}`.

This commit reworks the dictionary returned by `Pkg.Types.stdlibs()`
for Julia v1.8 and later, so that the rest of the existing code in
this package will work without modification.

Fixes #41.